pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teamflow"
version = "0.1.0"
description = "AI-powered project collaboration assistant for Feishu"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.22",
    "apscheduler>=3.10",
    "litellm>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "questionary>=2.1",
    "lark-oapi>=1.4.0",
]

[project.optional-dependencies]
setup = [
    "qrcode>=7.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "alembic>=1.13",
]

[project.scripts]
teamflow = "teamflow.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"src/teamflow/ai/model_registry.py" = ["E501"]
"src/teamflow/ai/skills/lark_cli.py" = ["E501"]
"src/teamflow/setup/cli.py" = ["E501"]
"src/teamflow/core/result.py" = ["UP046"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
